Move archive extraction into its own crate
