+++ crates/puffin-extract/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "puffin-extract"
+version = "0.0.1"
+edition = { workspace = true }
+rust-version = { workspace = true }
+homepage = { workspace = true }
+documentation = { workspace = true }
+repository = { workspace = true }
+authors = { workspace = true }
+license = { workspace = true }
+
+[lints]
+workspace = true
+
+[dependencies]
+flate2 = { workspace = true }
+fs-err = { workspace = true }
+rayon = { workspace = true }
+tar = { workspace = true }
+thiserror = { workspace = true }
+tokio = { workspace = true }
+zip = { workspace = true }
