+++ crates/puffin-installer/src/downloader.rs
@@ -1,6 +1,5 @@
 use std::cmp::Reverse;
 use std::path::PathBuf;
-
 use std::sync::Arc;
 
 use futures::{StreamExt, TryFutureExt};
@@ -12,18 +11,16 @@ use distribution_types::{CachedDist, Dist, RemoteSource, SourceDist};
 use platform_tags::Tags;
 use puffin_cache::Cache;
 use puffin_client::RegistryClient;
-use puffin_distribution::{
-    DistributionDatabase, DistributionDatabaseError, LocalWheel, Unzip, UnzipError,
-};
+use puffin_distribution::{DistributionDatabase, DistributionDatabaseError, LocalWheel, Unzip};
 use puffin_traits::{BuildContext, OnceMap};
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
     #[error("Failed to unzip wheel: {0}")]
-    Unzip(Dist, #[source] UnzipError),
+    Unzip(Dist, #[source] puffin_extract::Error),
     #[error("Failed to fetch wheel: {0}")]
     Fetch(Dist, #[source] DistributionDatabaseError),
-    /// Should not occur, i've only seen it when another task panicked
+    /// Should not occur; only seen when another task panicked.
     #[error("The task executor is broken, did some other task panic?")]
     Join(#[from] JoinError),
     #[error("Unzip failed in another thread: {0}")]
@@ -160,7 +157,7 @@ impl<'a, Context: BuildContext + Send + Sync> Downloader<'a, Context> {
 
         // Unzip the wheel.
         let normalized_path = tokio::task::spawn_blocking({
-            move || -> Result<PathBuf, UnzipError> {
+            move || -> Result<PathBuf, puffin_extract::Error> {
                 // Unzip the wheel into a temporary directory.
                 let parent = download
                     .target()
